requirements.txt
Flask==2.3.3
paho-mqtt==1.6.1
python-dotenv==1.0.0